# target/classes/swagger.yml
openapi: 3.0.3

info:
  title: md-user-service
  description: Microsserviço responsável por gerir todas as ações relacionadas a um usuário.
  version: 1.0.11

tags:
  - name: User

paths:
  /user:
    post:
      tags:
        - User
      summary: Create user
      operationId: createUser
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserModel'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseModel'
        '400':
          description: Bad request
        '404':
          description: Not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error

  /user/login:
    post:
      tags:
        - User
      summary: Login
      operationId: login
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginModel'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseModel'
        '400':
          description: Bad request
        '404':
          description: Not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error

  /user/profile/mentee:
    post:
      tags:
        - User
      summary: Complete mentee profile
      operationId: menteeProfile
      parameters:
        - $ref: '#/components/parameters/token'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenteeProfileModel'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseModel'
        '400':
          description: Bad request
        '404':
          description: Not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error

  /user/profile/mentor:
    post:
      tags:
        - User
      summary: Complete mentor profile
      operationId: mentorProfile
      parameters:
        - $ref: '#/components/parameters/token'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorProfileModel'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseModel'
        '400':
          description: Bad request
        '404':
          description: Not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error

  /user/mentee:
    get:
      tags:
        - User
      summary: Obtain a list of mentees available for sponsorship
      operationId: getMenteesList
      parameters:
        - $ref: '#/components/parameters/token'
        - $ref: '#/components/parameters/limitQuery'
        - $ref: '#/components/parameters/pageQuery'
        - $ref: '#/components/parameters/email'
      responses:
        '200':
          description: List of mentees successfully returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMentees'
        '400':
          description: Bad request
        '404':
          description: Not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error

  /user/password/forgot:
    post:
      tags:
        - User
      summary: Request password reset link
      operationId: forgotPassword
      parameters:
        - $ref: '#/components/parameters/token'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordModel'
      responses:
        '200':
          description: Password recovery request sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseModel'
        '400':
          description: Bad request
        '404':
          description: Not found
        '422':
          description: Unprocessable Entity
        '500':
          description: Internal Server Error


components:
  parameters:
    token:
      name: token
      in: header
      required: true
      description: JWT User Access Token
      schema:
        type: string

    email:
      name: email
      in: header
      required: true
      description: User email
      schema:
        type: string

    limitQuery:
      name: limit
      in: query
      required: false
      description: Number of records per page
      schema:
        type: integer
        example: 2
        minimum: 2
        default: 2

    pageQuery:
      name: page
      in: query
      required: false
      description: Current page
      schema:
        type: integer
        example: 1
        minimum: 1
        default: 1

  schemas:
    UserModel:
      type: object
      properties:
        name:
          type: string
          example: Maria Eugênia Nascimento
        email:
          type: string
          example: mariaeugenia@email.com
        password:
          type: string
          example: m4ria3ug3ni4
        phone:
          type: string
          example: 31995236485
        image:
          type: string
        isMentor:
          enum:
            - 1
            - 2

    LoginModel:
      properties:
        email:
          type: string
          example: mariaeugenia@email.com
        password:
          type: string
          example: m4ria3ug3ni4

    MenteeProfileModel:
      properties:
        email:
          type: string
          example: mariaeugenia@email.com
        profile:
          type: string
          example: Olá, sou a Maria Eugênia [...]
        isSponsored:
          type: boolean
          default: false
        age:
          type: string
          example: 24
        menteeLevel:
          enum:
            - 1
            - 2

    MentorProfileModel:
      properties:
        email:
          type: string
          example: mariaeugenia@email.com
        education:
          type: string
          example: Bacharelado em Sistemas de Informação
        age:
          type: string
          example: 24
        mentoringCapacity:
          enum:
            - 1
            - 2
          default: 1
        profile:
          type: string
          example: Olá, sou a Maria Eugênia [...]

    MenteesModel:
      type: object
      properties:
        name:
          type: string
          example: Maria Eugênia Nascimento
        email:
          type: string
          example: mariaeugenia@email.com
        phone:
          type: string
          example: 31995236485
        profile:
          type: string
          example: Olá, sou a Maria Eugênia [...]
        age:
          type: string
          example: 24
        menteeLevel:
          type: string
          example: LAGARTA
        image:
          type: string
          format: binary

    PaginatedMentees:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MenteesModel'
        currentPage:
          type: integer
          example: 1
        totalRecords:
          type: integer
          example: 1
        totalRecordsPerPage:
          type: integer
          example: 2

    ResponseModel:
      type: object
      properties:
        mensagem:
          type: string
        status:
          type: number

    ForgotPasswordModel:
      type: object
      properties:
        email:
          type: string

    LoginResponseModel:
      type: object
      properties:
        mensagem:
          type: string
        status:
          type: number
        token:
          type: string