# target/generated-sources/openapi/src/main/java/mariposas/model/UserModel.java
/*
 * md-user-service
 * Microsserviço responsável por gerir todas as ações relacionadas a um usuário.
 *
 * The version of the OpenAPI document: 1.0.11
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package mariposas.model;

import java.util.Objects;
import java.util.Arrays;
import java.util.Map;
import java.util.stream.Collectors;
import mariposas.model.UserModelIsMentor;
import com.fasterxml.jackson.annotation.*;
import io.micronaut.serde.annotation.Serdeable;
import jakarta.validation.constraints.*;
import jakarta.validation.Valid;
import io.micronaut.core.annotation.Nullable;
import jakarta.annotation.Generated;
import io.swagger.v3.oas.annotations.media.Schema;

/**
 * UserModel
 */
@Serdeable
@JsonPropertyOrder({
    UserModel.JSON_PROPERTY_NAME,
    UserModel.JSON_PROPERTY_EMAIL,
    UserModel.JSON_PROPERTY_PASSWORD,
    UserModel.JSON_PROPERTY_PHONE,
    UserModel.JSON_PROPERTY_IMAGE,
    UserModel.JSON_PROPERTY_IS_MENTOR
})
@Generated("io.micronaut.openapi.generator.JavaMicronautServerCodegen")
public class UserModel {

    public static final String JSON_PROPERTY_NAME = "name";
    public static final String JSON_PROPERTY_EMAIL = "email";
    public static final String JSON_PROPERTY_PASSWORD = "password";
    public static final String JSON_PROPERTY_PHONE = "phone";
    public static final String JSON_PROPERTY_IMAGE = "image";
    public static final String JSON_PROPERTY_IS_MENTOR = "isMentor";

    @Nullable(inherited = true)
    @Schema(name = "name", example = "Maria Eugênia Nascimento", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
    @JsonProperty(JSON_PROPERTY_NAME)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private String name;

    @Nullable(inherited = true)
    @Schema(name = "email", example = "mariaeugenia@email.com", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
    @JsonProperty(JSON_PROPERTY_EMAIL)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private String email;

    @Nullable(inherited = true)
    @Schema(name = "password", example = "m4ria3ug3ni4", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
    @JsonProperty(JSON_PROPERTY_PASSWORD)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private String password;

    @Nullable(inherited = true)
    @Schema(name = "phone", example = "31995236485", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
    @JsonProperty(JSON_PROPERTY_PHONE)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private String phone;

    @Nullable(inherited = true)
    @Schema(name = "image", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
    @JsonProperty(JSON_PROPERTY_IMAGE)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private String image;

    @Nullable(inherited = true)
    @Schema(name = "isMentor", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
    @JsonProperty(JSON_PROPERTY_IS_MENTOR)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private UserModelIsMentor isMentor;

    /**
     * @return the name property value
     */
    public String getName() {
        return name;
    }

    /**
     * Set the name property value
     */
    public void setName(String name) {
        this.name = name;
    }

    /**
     * Set name in a chainable fashion.
     *
     * @return The same instance of UserModel for chaining.
     */
    public UserModel name(String name) {
        this.name = name;
        return this;
    }

    /**
     * @return the email property value
     */
    public String getEmail() {
        return email;
    }

    /**
     * Set the email property value
     */
    public void setEmail(String email) {
        this.email = email;
    }

    /**
     * Set email in a chainable fashion.
     *
     * @return The same instance of UserModel for chaining.
     */
    public UserModel email(String email) {
        this.email = email;
        return this;
    }

    /**
     * @return the password property value
     */
    public String getPassword() {
        return password;
    }

    /**
     * Set the password property value
     */
    public void setPassword(String password) {
        this.password = password;
    }

    /**
     * Set password in a chainable fashion.
     *
     * @return The same instance of UserModel for chaining.
     */
    public UserModel password(String password) {
        this.password = password;
        return this;
    }

    /**
     * @return the phone property value
     */
    public String getPhone() {
        return phone;
    }

    /**
     * Set the phone property value
     */
    public void setPhone(String phone) {
        this.phone = phone;
    }

    /**
     * Set phone in a chainable fashion.
     *
     * @return The same instance of UserModel for chaining.
     */
    public UserModel phone(String phone) {
        this.phone = phone;
        return this;
    }

    /**
     * @return the image property value
     */
    public String getImage() {
        return image;
    }

    /**
     * Set the image property value
     */
    public void setImage(String image) {
        this.image = image;
    }

    /**
     * Set image in a chainable fashion.
     *
     * @return The same instance of UserModel for chaining.
     */
    public UserModel image(String image) {
        this.image = image;
        return this;
    }

    /**
     * @return the isMentor property value
     */
    public UserModelIsMentor getIsMentor() {
        return isMentor;
    }

    /**
     * Set the isMentor property value
     */
    public void setIsMentor(UserModelIsMentor isMentor) {
        this.isMentor = isMentor;
    }

    /**
     * Set isMentor in a chainable fashion.
     *
     * @return The same instance of UserModel for chaining.
     */
    public UserModel isMentor(UserModelIsMentor isMentor) {
        this.isMentor = isMentor;
        return this;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        UserModel userModel = (UserModel) o;
        return Objects.equals(name, userModel.name) &&
            Objects.equals(email, userModel.email) &&
            Objects.equals(password, userModel.password) &&
            Objects.equals(phone, userModel.phone) &&
            Objects.equals(image, userModel.image) &&
            Objects.equals(isMentor, userModel.isMentor);
    }

    @Override
    public int hashCode() {
        return Objects.hash(name, email, password, phone, image, isMentor);
    }

    @Override
    public String toString() {
        return "UserModel("
            + "name: " + getName() + ", "
            + "email: " + getEmail() + ", "
            + "password: " + getPassword() + ", "
            + "phone: " + getPhone() + ", "
            + "image: " + getImage() + ", "
            + "isMentor: " + getIsMentor()
            + ")";
    }

}